Cleanup eslint warnings, add public board url with input
