+++ src/components/public/PublicColumn.jsx
@@ -1,51 +1,8 @@
 import { Droppable, Draggable } from 'react-beautiful-dnd'
-import Task from './Task'
-
-import {Bin, Exclaim} from './Icons'
-
-import {db, firebase} from '../firebase/fbConfig'
-import {debounce} from '../utils'
-import {useState, useRef} from 'react'
-import Modal from './Modal'
 import PublicTask from './PublicTask'
 
 
 const PublicColumn = ({ column, tasks, allData, boardId, userId, filterBy, index }) => {
-
-    const [modal, setModal] = useState(false)
-    const [editingCol, setEditing] = useState(false)
-    const colInput = useRef(null)
-
-    const deleteCol = (colId, tasks) => {
-        db.collection(`users/${userId}/boards/${boardId}/columns`)        
-            .doc('columnOrder')
-            .update({order: firebase.firestore.FieldValue.arrayRemove(colId)})   
-
-        db.collection(`users/${userId}/boards/${boardId}/columns`)
-            .doc(colId)
-            .delete()
-
-        //Extract and delete its tasks
-        tasks.forEach(t => {
-            db.collection(`users/${userId}/boards/${boardId}/tasks`)
-            .doc(t)
-            .delete()
-        })
-    }
-
-    const changeColName = debounce((e, colId) => {
-        db.collection(`users/${userId}/boards/${boardId}/columns`)
-            .doc(colId)
-            .update({title: e.target.value})
-    }, 7000)
-
-    const moveToInp = () => {
-        setEditing(true)
-        setTimeout(()=>{
-            colInput.current.focus()
-        }, 50)
-    }
-
 
     return (
         <>
