+++ src/screens/Publish.jsx
@@ -1,20 +1,26 @@
 
 import {useState} from 'react'
-import {db, firebase} from '../firebase/fbConfig'
-import {v4 as uuidv4} from 'uuid';
+import { Copy } from '../components/Icons';
 import Toggle from '../components/Toggle';
 import { copyBoardsToPublicBoards } from '../utils';
 
 const Publish = ({ close, userId, boardId, boardName, data }) => 
 {
-	const [publish,setPublish] = useState(true);
+	const [publish,setPublish] = useState(false);
+	const publicBoardUrl = window.location.href.replace('board/','public-board/');
 
+	const copyUrl = () =>{
+		const textBox = document.querySelector(".public-url");
+		textBox.select();
+		document.execCommand("copy");
+	}
 	const publishBoard = async (status)=>{
-		 if(!status){
+		 if(status){
 			 setPublish(false);
 			 return;
 			}
-		copyBoardsToPublicBoards(userId, boardId, boardName, data );
+			setPublish(true);
+			// copyBoardsToPublicBoards(userId, boardId, boardName, data );
 	}
 
 	return (
@@ -22,8 +28,12 @@ const Publish = ({ close, userId, boardId, boardName, data }) =>
 			<h2>Publish this board.</h2>
 			<div className="flex my-2">
 				<Toggle status={publishBoard} />
-				<span className="ml-2">{publish ? 'Publish' : 'Published'}</span>
+				<span className="ml-2">{ publish ? 'Published' : 'Publish'}</span>
 			</div>
+			{ publish ? ( <div className="flex">
+			 <input className="public-url p-2 border" style={{maxWidth:'500px',width:'100%'}} type="text" defaultValue={publicBoardUrl} />
+			<button onClick={copyUrl} className="p-2 border-none outline-none"><Copy /></button>
+			</div> ) : '' }
 
 		</div>
 	)
